Remove unused code and fix conflicts
