+++ libs/remix-ui/workspace/src/lib/components/file-explorer.tsx
@@ -2,9 +2,7 @@ import React, { useEffect, useState, useRef, SyntheticEvent } from 'react' // es
 import { TreeView, TreeViewItem } from '@remix-ui/tree-view' // eslint-disable-line
 import { FileExplorerMenu } from './file-explorer-menu' // eslint-disable-line
 import { FileExplorerContextMenu } from './file-explorer-context-menu' // eslint-disable-line
-import { FileExplorerProps, MenuItems, FileExplorerState, WorkSpaceState } from '../types'
-import { customAction } from '@remixproject/plugin-api'
-import { contextMenuActions } from '../utils'
+import { FileExplorerProps, WorkSpaceState } from '../types'
 
 import '../css/file-explorer.css'
 import { checkSpecialChars, extractNameFromKey, extractParentFromKey, joinPath } from '@remix-ui/helper'
@@ -15,7 +13,7 @@ import { ROOT_PATH } from '../utils/constants'
 
 export const FileExplorer = (props: FileExplorerProps) => {
   const { name, contextMenuItems, removedContextMenuItems, files, workspaceState, toGist, addMenuItems, 
-    removeMenuItems, handleContextMenu, handleNewFileInput, handleNewFolderInput, uploadFile, fileState } = props
+    removeMenuItems, handleContextMenu, handleNewFileInput, handleNewFolderInput, uploadFile, uploadFolder, fileState } = props
   const [state, setState] = useState<WorkSpaceState>( workspaceState)
   const treeRef = useRef<HTMLDivElement>(null)
   
@@ -97,46 +95,6 @@ export const FileExplorer = (props: FileExplorerProps) => {
       props.dispatchRenamePath(oldPath, newPath)
     } catch (error) {
       props.modal('Rename File Failed', 'Unexpected error while renaming: ' + typeof error === 'string' ? error : error.message, 'Close', async () => {})
-    }
-  }
-
-  const downloadPath = async (path: string) => {
-    try {
-      props.dispatchDownloadPath(path)
-    } catch (error) {
-      props.modal('Download Failed', 'Unexpected error while downloading: ' + typeof error === 'string' ? error : error.message, 'Close', async () => {})
-    }
-  }
-
-  const uploadFile = (target) => {
-    const parentFolder = getFocusedFolder()
-    const expandPath = [...new Set([...props.expandPath, parentFolder])]
-
-    props.dispatchHandleExpandPath(expandPath)
-    props.dispatchUploadFile(target, parentFolder)
-  }
-
-  const uploadFolder = (target) => {
-    const parentFolder = getFocusedFolder()
-    const expandPath = [...new Set([...props.expandPath, parentFolder])]
-
-    props.dispatchHandleExpandPath(expandPath)
-    props.dispatchUploadFolder(target, parentFolder)
-  }
-
-  const copyFile = (src: string, dest: string) => {
-    try {
-      props.dispatchCopyFile(src, dest)
-    } catch (error) {
-      props.modal('Copy File Failed', 'Unexpected error while copying file: ' + src, 'Close', async () => {})
-    }
-  }
-
-  const copyFolder = (src: string, dest: string) => {
-    try {
-      props.dispatchCopyFolder(src, dest)
-    } catch (error) {
-      props.modal('Copy Folder Failed', 'Unexpected error while copying folder: ' + src, 'Close', async () => {})
     }
   }
 
@@ -274,76 +232,50 @@ export const FileExplorer = (props: FileExplorerProps) => {
 
   return (
     <Drag onFileMoved={handleFileMove} onFolderMoved={handleFolderMove}>
-      <div ref={treeRef} tabIndex={0} style={{ outline: "none" }}>
-        <TreeView id='treeView'>
-          <TreeViewItem id="treeViewItem"
-            controlBehaviour={true}
-            label={
-              <div onClick={handleFileExplorerMenuClick}>
-                <FileExplorerMenu
-                  title={''}
-                  menuItems={props.menuItems}
-                  createNewFile={handleNewFileInput}
-                  createNewFolder={handleNewFolderInput}
-                  publishToGist={publishToGist}
-                  uploadFile={uploadFile}
-                  uploadFolder={uploadFolder}
-                />
-              </div>
-            }
-            expand={true}>
-            <div className='pb-4 mb-4'>
-              <TreeView id='treeViewMenu'>
-                {
-                  files[ROOT_PATH] && Object.keys(files[ROOT_PATH]).map((key, index) => <FileRender
-                    file={files[ROOT_PATH][key]}
-                    fileDecorations={fileState}
-                    index={index}
-                    focusContext={state.focusContext}
-                    focusEdit={state.focusEdit}
-                    focusElement={props.focusElement}
-                    ctrlKey={state.ctrlKey}
-                    expandPath={props.expandPath}
-                    editModeOff={editModeOff}
-                    handleClickFile={handleClickFile}
-                    handleClickFolder={handleClickFolder}
-                    handleContextMenu={handleContextMenu}
-                    key={index}
-                    showIconsMenu={props.showIconsMenu}
-                    hideIconsMenu={props.hideIconsMenu}
-                    
-                  />)
-                }
-              </TreeView>
+    <div ref={treeRef} tabIndex={0} style={{ outline: "none" }}>
+      <TreeView id='treeView'>
+        <TreeViewItem id="treeViewItem"
+          controlBehaviour={true}
+          label={
+            <div onClick={handleFileExplorerMenuClick}>
+              <FileExplorerMenu
+                title={''}
+                menuItems={props.menuItems}
+                createNewFile={handleNewFileInput}
+                createNewFolder={handleNewFolderInput}
+                publishToGist={publishToGist}
+                uploadFile={uploadFile}
+                uploadFolder={uploadFolder}
+              />
             </div>
-          </TreeViewItem>
-        </TreeView>
-        { state.showContextMenu &&
-          <FileExplorerContextMenu
-            actions={props.focusElement.length > 1 ? state.actions.filter(item => item.multiselect) : state.actions.filter(item => !item.multiselect)}
-            hideContextMenu={hideContextMenu}
-            createNewFile={handleNewFileInput}
-            createNewFolder={handleNewFolderInput}
-            deletePath={deletePath}
-            downloadPath={downloadPath}
-            renamePath={editModeOn}
-            runScript={runScript}
-            copy={handleCopyClick}
-            paste={handlePasteClick}
-            copyFileName={handleCopyFileNameClick}
-            copyPath={handleCopyFilePathClick}
-            emit={emitContextMenuEvent}
-            pageX={state.focusContext.x}
-            pageY={state.focusContext.y}
-            path={state.focusContext.element}
-            type={state.focusContext.type}
-            focus={props.focusElement}
-            pushChangesToGist={pushChangesToGist}
-            publishFolderToGist={publishFolderToGist}
-            publishFileToGist={publishFileToGist}
-          />
-        }
-      </div>
+          }
+          expand={true}>
+          <div className='pb-4 mb-4'>
+            <TreeView id='treeViewMenu'>
+              {
+                files[ROOT_PATH] && Object.keys(files[ROOT_PATH]).map((key, index) => <FileRender
+                file={files[ROOT_PATH][key]}
+                fileDecorations={fileState}
+                index={index}
+                focusContext={state.focusContext}
+                focusEdit={state.focusEdit}
+                focusElement={props.focusElement}
+                ctrlKey={state.ctrlKey}
+                expandPath={props.expandPath}
+                editModeOff={editModeOff}
+                handleClickFile={handleClickFile}
+                handleClickFolder={handleClickFolder}
+                handleContextMenu={handleContextMenu}
+                key={index}
+                showIconsMenu={props.showIconsMenu}
+                hideIconsMenu={props.hideIconsMenu}
+              />)
+              }
+            </TreeView>
+          </div>
+        </TreeViewItem>
+      </TreeView>
+    </div>
     </Drag>
   )
 }
